+++ Utopia/Classes/Battle/BattleOrbLayout.m
@@ -142,7 +142,7 @@
   return set;
 }
 
-- (void) generateRandomOrbColor:(OrbColor *)orbColor specialOrbType:(SpecialOrbType *)specialOrbType atColumn:(int)column row:(int)row {
+- (void) generateRandomOrbColor:(OrbColor *)orbColor specialOrbType:(SpecialOrbType *)specialOrbType atColumn:(NSInteger)column row:(NSInteger)row {
   
   SpecialOrbType orbType = SpecialOrbTypeNone;
   if (row > _numRows/2)
@@ -589,7 +589,7 @@
   BOOL isHorizontal = lineOrb.powerupType == PowerupTypeHorizontalLine;
   
   if (isHorizontal) {
-    for (int i = swap.orbA.row-1; i <= swap.orbA.row+1; i++) {
+    for (NSInteger i = swap.orbA.row-1; i <= swap.orbA.row+1; i++) {
       if (i >= 0 && i < _numRows) {
         BattleOrb *orb = swap.orbA.copy;
         orb.row = i;
@@ -604,7 +604,7 @@
       }
     }
   } else {
-    for (int i = swap.orbA.column-1; i <= swap.orbA.column+1; i++) {
+    for (NSInteger i = swap.orbA.column-1; i <= swap.orbA.column+1; i++) {
       if (i >= 0 && i < _numColumns) {
         BattleOrb *orb = swap.orbA.copy;
         orb.column = i;
@@ -967,8 +967,8 @@
   chain.prerequisiteOrb = orb;
   chain.chainType = ChainTypePowerupNormal;
   
-  for (int i = orb.column-1; i <= orb.column+1; i++) {
-    for (int j = orb.row-1; j <= orb.row+1; j++) {
+  for (NSInteger i = orb.column-1; i <= orb.column+1; i++) {
+    for (NSInteger j = orb.row-1; j <= orb.row+1; j++) {
       if (i >= 0 && i < _numColumns && j >= 0 && j < _numRows) {
         BattleOrb *testOrb = [self orbAtColumn:i row:j];
         if ([self orbCanBeRemoved:testOrb]) {
